Resolved: TODO(roasbeef): similar interface for headerfs? introduced BlockHeaderDB and FilterHeaderDB interfaces that expose the previously exported functions.
